add completed version of node chat app

+++ src/services/message.service.js
@@ -4,6 +4,7 @@ const { Message } = require('../models/Message');
 const userService = require('./user.service');
 const roomService = require('./room.service');
 const directService = require('./direct.service');
+const { ApiError } = require('../exceptions/ApiError');
 
 const normalize = ({ id, author, text, createdAt }) => {
   return {
@@ -15,39 +16,62 @@ const normalize = ({ id, author, text, createdAt }) => {
 };
 
 const getById = (messageId) => {
-  return Message.findOne({ where: { id: messageId } });
+  return Message.findByPk(messageId);
 };
 
-const getAllByRoom = (roomId) => {
-  return Message.findAll({ where: { roomId } });
+const getAllByRoom = async (roomId) => {
+  if (!(await roomService.getById(roomId))) {
+    throw ApiError.NotFound('Room does not exist');
+  }
+
+  return Message.findAll({ where: { roomId }, order: [['createdAt', 'ASC']] });
 };
 
-const deleteAllByRoom = (roomId) => {
+const deleteAllByRoom = async (roomId) => {
+  if (!(await roomService.getById(roomId))) {
+    throw ApiError.NotFound('Room does not exist');
+  }
+
   return Message.destroy({ where: { roomId } });
 };
 
-const getAllByDirect = (directId) => {
-  return Message.findAll({ where: { directId } });
+const deleteAllByDirect = async (directId) => {
+  if (!(await directService.getById(directId))) {
+    throw ApiError.NotFound('Direct does not exist');
+  }
+
+  return Message.destroy({ where: { directId } });
+};
+
+const getAllByDirect = async (directId) => {
+  if (!(await directService.getById(directId))) {
+    throw ApiError.NotFound('Dialogue does not exist');
+  }
+
+  return Message.findAll({
+    where: { directId },
+    order: [['createdAt', 'ASC']],
+  });
 };
 
 const add = async ({ text, author, roomId, directId }) => {
   const user = await userService.getByName(author);
 
   if (!user) {
-    throw new Error('User does not exist');
+    throw new ApiError.NotFound('User does not exist');
   }
 
   if (roomId) {
     const room = await roomService.getById(roomId);
 
     if (!room) {
-      throw new Error('Room does not exist');
+      throw new ApiError.NotFound('Room does not exist');
     }
   } else if (directId) {
     const direct = await directService.getById(directId);
 
     if (!direct) {
-      throw new Error('Direct does not exist');
+      throw new ApiError.NotFound('Direct does not exist');
     }
   }
 
@@ -59,11 +83,35 @@ const add = async ({ text, author, roomId, directId }) => {
   });
 };
 
+const edit = async ({ messageId, newText }) => {
+  const message = await getById(messageId);
+
+  if (!message) {
+    throw new ApiError.NotFound('Message not found');
+  }
+
+  message.text = newText;
+  await message.save();
+};
+
+const remove = async (messageId) => {
+  const message = await getById(messageId);
+
+  if (!message) {
+    throw new ApiError.NotFound('Message not found');
+  }
+
+  await Message.destroy({ where: { id: messageId } });
+};
+
 module.exports = {
   normalize,
   getById,
   getAllByRoom,
   getAllByDirect,
   deleteAllByRoom,
+  deleteAllByDirect,
   add,
+  edit,
+  remove,
 };
